Split the main network queue out of MainQueue so libs/Log can queue without importing libs/Network
libs/Log imported `post` from libs/Network, which imports MainQueue, which imports libs/Request,
SequentialQueue and NetworkStore, all three of which import libs/Log. That one edge held the 15-file
network cluster inside oxlint's import/no-cycle graph.

Every log packet is queued with shouldProcessImmediately: false, so `post` pushed it and returned
without calling MainQueue.process. So the queue and the function that puts a request on it move to
Network/MainQueueStore, which imports nothing that reaches libs/Log. MainQueue re-exports push, clear
and getAll, so its consumers do not change, and libs/Log imports `enqueue` rather than `post`.

No function body changes. `post` is `enqueue` plus the process kick it already made conditionally.

import/no-cycle goes from 534 findings across 148 files to 500 across 133, and the libs/Network and
libs/Request subsystem leaves the cycle graph. No other oxlint rule count changes.
